feat: schedule backups daily (#4)
Currently, we just run the backups on a 24 hour (ish) schedule from the
time the container starts. This isn't great, since database traffic
might be higher at certain points of the day. Running them overnight is
probably a better idea, so let's use `Interval` to make this happen.

This change:
* Updates the loop to use `Interval` and run at 22:30 daily
* Adds some tests for the scheduling logic